// modules/players/interfaces/Player.ts
export interface Player {
  id: number;
  code: number;
  team: number;
  webName: string;
  firstName: string;
  secondName: string;
  goalsScored: number;
  assists: number;
  cleanSheets: number;
  redCards: number;
  price: string;
  news: string;
  position: number;
  image: string;
  imageLarge: string;
  teamName: string;
  teamNameShort: string;
  status: string;
  isUnavailable: boolean;
  unavailableForSeason: boolean;
}